+++ HotelWise.Domain/Interfaces/IA/IAIInferenceAdapter.cs
@@ -2,7 +2,7 @@
 {
     public interface IAIInferenceAdapter
     {
-        Task<decimal[]> GenerateEmbeddingAsync(string text);
+        Task<float[]> GenerateEmbeddingAsync(string text);
         Task<string> GenerateChatCompletionAsync(string prompt);
     }
 }
